bank16kfc:  fix some types to get nicer code

+++ Kernel/bank16kfc.c
@@ -82,8 +82,8 @@ static int maps_needed(uint16_t top)
 int pagemap_alloc(ptptr p)
 {
 	uint8_t *ptr = (uint8_t *) & p->p_page;
-	int needed = maps_needed(p->p_top);
-	int i;
+	uint_fast8_t needed = maps_needed(p->p_top);
+	uint_fast8_t i;
 
 #ifdef SWAPDEV
 	/* Throw our toys out of our pram until we have enough room */
